refactor: rename codegen module to compile
