+++ crates/gantz_core/src/compile/node_fns.rs
