+++ crates/gantz_core/src/lib.rs
@@ -2,7 +2,7 @@ pub use edge::Edge;
 pub use node::Node;
 pub use steel;
 
-pub mod codegen;
+pub mod compile;
 pub mod edge;
 pub mod graph;
 pub mod node;
