+++ crates/gantz_core/tests/state.rs
@@ -2,7 +2,7 @@
 
 use gantz_core::{
     Edge, ROOT_STATE,
-    codegen::push_eval_fn_name,
+    compile::push_eval_fn_name,
     node::{self, Node, NodeState, WithPushEval, WithStateType},
 };
 use std::fmt::Debug;
@@ -82,7 +82,7 @@ fn test_graph_with_counter() {
     g.add_edge(push, counter, Edge::from((0, 0)));
 
     // Generate the module, which should have just one top-level expr for `push`.
-    let module = gantz_core::codegen::module(&g);
+    let module = gantz_core::compile::module(&g);
 
     // Initialise the VM.
     let mut vm = Engine::new_base();
@@ -172,7 +172,7 @@ fn test_graph_with_counters() {
     g.add_edge(p_c, c_c, Edge::from((0, 0)));
 
     // Generate the module, which should have one expr for each `push`.
-    let module = gantz_core::codegen::module(&g);
+    let module = gantz_core::compile::module(&g);
 
     // Initialise the VM.
     let mut vm = Engine::new_base();
